--- iree/compiler/.gitignore
@@ -1,1 +1,0 @@
-*.whl
