+++ iree/compiler/API/Tools.cpp
@@ -4,6 +4,10 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Api/Tools.h"
+#include "iree/compiler/API/Tools.h"
 
-int main(int argc, char **argv) { return ireeCompilerRunMain(argc, argv); }
+#include "iree/tools/iree_translate_lib.h"
+
+int ireeCompilerRunMain(int argc, char **argv) {
+  return mlir::iree_compiler::runIreecMain(argc, argv);
+}
