+++ iree/compiler/API/python/test/tools/testdata/tflite_sample.fb
